requirements.txt
django>=4.0.4
django-debug-toolbar==3.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
whitenoise==6.1.0
djangorestframework==3.13.1
